Simplify process.
Redeclare variables related to user input and number of panels. Store all processes related to variable assignment and operations inside a function with a condition that does not allow for a bigger input than 100.

Replace variable re-assignment declarations on change button and replace with function in charge of doing so.